requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
email-validator
python-jose[cryptography]
